pyproject.toml
[tool.poetry]
name = "congressmoney"
version = "0.1.0"
description = "An open-source platform to promote financial transparency by making the stock trades and personal finances broadly accessible."
authors = ["George McCain <georgemccain88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
requests = "^2.26.0"
python-dotenv = "^0.18.0"
black = "^21.7b0"
python-dateutil = "^2.8.2"
diskcache = "^5.2.1"
tqdm = "^4.62.0"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
